CSS Properties can now be indented
This indentation is controlled by the variable *indent-css*.

The default value of this variable is NIL. This indicates that no
indentation should be performed, thereby conserving the old behaviour
of css-lite.

If the value of *indent-css* is 'tab, then the properties are indented
using a #\Tab character.

If the value of *indent-css* is a integer greater than 0, then the
properties are indented using that many #\Space characters.