CSS comments no longer add the extra ':' and ';'
Modified the function process-css-properties so it can process the
comments in a special way.

In other words this means that comments are no longer printed in this
way:

/*a comment:*/;

but instead are printed in this format:

/*a comment*/

Also changed the output of some examples in the file
example-usage.lisp to match the this new output.